+++ src/getStepStartStates.test.ts
@@ -80,7 +80,10 @@ import {
   STEP_RM_BATCH_APPLICATION,
   STEP_RM_BATCH_CERTIFICATE,
 } from './steps/resource-manager/batch';
-import { STEP_RM_REDIS_CACHES } from './steps/resource-manager/redis-cache';
+import {
+  STEP_RM_REDIS_CACHES,
+  STEP_RM_REDIS_FIREWALL_RULES,
+} from './steps/resource-manager/redis-cache';
 import {
   STEP_RM_EVENT_GRID_DOMAINS,
   STEP_RM_EVENT_GRID_DOMAIN_TOPICS,
@@ -155,6 +158,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_APPLICATION]: { disabled: true },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
       [STEP_RM_REDIS_CACHES]: { disabled: true },
+      [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -223,6 +227,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_APPLICATION]: { disabled: true },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: true },
       [STEP_RM_REDIS_CACHES]: { disabled: true },
+      [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: true },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: true },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: true },
@@ -291,6 +296,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_BATCH_APPLICATION]: { disabled: false },
       [STEP_RM_BATCH_CERTIFICATE]: { disabled: false },
       [STEP_RM_REDIS_CACHES]: { disabled: false },
+      [STEP_RM_REDIS_FIREWALL_RULES]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAINS]: { disabled: false },
       [STEP_RM_EVENT_GRID_DOMAIN_TOPICS]: { disabled: false },
       [STEP_RM_EVENT_GRID_TOPICS]: { disabled: false },