+++ CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to
 - Added `azure_batch_account|has|azure_batch_certificate` relationships
 - Added `azure_redis_cache` entities
 - Added `azure_resource_group|has|azure_redis_cache` relationships
+- Added `azure_redis_firewall_rule` entities
+- Added `azure_redis_cache|has|azure_redis_firewall_rule` relationships
 
 ## 5.3.1 - 2020-09-21